Move processors under subfolder
