+++ packages/consensus/source/index.ts
@@ -6,9 +6,9 @@ import { Aggregator } from "./aggregator";
 import { Bootstrapper } from "./bootstrapper";
 import { CommittedBlockProcessor } from "./committed-block-processor";
 import { Consensus } from "./consensus";
-import { PrecommitProcessor } from "./precommit-processor";
-import { PrevoteProcessor } from "./prevote-processor";
-import { ProposalProcessor } from "./proposal-processor";
+import { PrecommitProcessor } from "./processors/precommit-processor";
+import { PrevoteProcessor } from "./processors/prevote-processor";
+import { ProposalProcessor } from "./processors/proposal-processor";
 import { ProposerPicker } from "./proposer-picker";
 import { RoundStateRepository } from "./round-state-repository";
 import { Scheduler } from "./scheduler";
